refactor: move postUser to api file
