+++ client/api/index.ts
@@ -1,4 +1,5 @@
 import { User } from "components/UserList";
+import { v4 as uuidv4 } from "uuid";
 const baseURL = process.env.NEXT_PUBLIC_API_BASE_URL;
 
 async function listUsers() {
@@ -8,4 +9,22 @@ async function listUsers() {
   return parsed;
 }
 
-export { listUsers };
+async function postUser(username: string, name: string) {
+  // TODO: remove this once the back end handles id creation
+  const id = uuidv4();
+  const url = `${baseURL}/users`;
+  const res = await fetch(url, {
+    method: "POST",
+    headers: {
+      Accept: "application/json",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ username, name, id }),
+  });
+  if (res.ok) {
+    // TODO: make this depend on the server response?
+    return `User ${username} was successfully created`;
+  }
+}
+
+export { listUsers, postUser };
