+++ client/components/UserList.tsx
@@ -1,6 +1,5 @@
 import { FormEvent, useEffect, useRef, useState } from "react";
-import { v4 as uuidv4 } from "uuid";
-import { listUsers } from "api";
+import { listUsers, postUser } from "api";
 
 export interface User {
   username: string;
@@ -34,21 +33,9 @@ export default function UserList() {
   }
 
   async function createUser(username: string, name: string) {
-    const url = `${process.env.NEXT_PUBLIC_API_BASE_URL}/users`;
-    // TODO: remove this once the back end handles id creation
-    const id = uuidv4();
     try {
-      const response = await fetch(url, {
-        method: "POST",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ username, name, id }),
-      });
-      if (response.ok) {
-        setStatusMessage(`User ${username} successfully created`);
-      }
+      const response = await postUser(username, name);
+      setStatusMessage(response);
     } catch (err) {
       console.error(err);
     }
