+++ client/pages/index.tsx
@@ -4,7 +4,7 @@ import ExpenseList from "components/ExpenseList";
 
 export default function Home() {
   return (
-    <div className="container mx-auto px-4">
+    <div className="container mx-auto px-4 py-8">
       <Head>
         <title>Expenseus</title>
         <meta name="description" content="Generated by create next app" />
